Remove insignificant nested directories
