+++ packages/core/src/lib/chains/algorand/transactions/tokenBridge/index.ts
@@ -6,7 +6,7 @@ import {
     Routing,
     RoutingDefault,
 } from "src/lib/common";
-import {BridgeNetworks} from "src/lib/common/networks/networks";
+import {BridgeNetworks} from "src/lib/common/networks";
 import {
     algoTransferTxnWithRoutingNote,
     assetTransferTxnWithRoutingNote,
