+++ packages/core/src/lib/configs/index.ts
@@ -1,0 +1,5 @@
+import {config as mConfig} from "./mainnet"
+import {config as tConfig} from "./testnet"
+
+export const mainnetConfig = mConfig;
+export const testnetConfig = tConfig;
